Add on the go theme switching
